+++ src/ElasticsearchMemoryStorage/ElasticsearchMemoryRecord.cs
@@ -1,7 +1,5 @@
 ﻿// Copyright (c) Free Mind Labs, Inc. All rights reserved.
 
-using System;
-using System.Collections.Generic;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using Microsoft.KernelMemory;
