+++ src/ElasticsearchMemoryStorage/Extensions/KernelMemoryBuilderExtensions.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) Free Mind Labs, Inc. All rights reserved.
 
-using System;
 using Microsoft.Extensions.Configuration;
 using Microsoft.KernelMemory;
